+++ feature/assist/src/main/java/com/linnan/blindassist/runtime/AssistRuntimeRenderer.kt
@@ -21,6 +21,10 @@ internal class AssistRuntimeRenderer(
         overlayView = overlay
     }
 
+    fun detachOverlay() {
+        overlayView = null
+    }
+
     fun currentOverlay(): DetectionOverlayView? = overlayView
 
     fun renderInitial() {